🎨 Partition report into stages
👷 Push book on change to `config`, `res`

📝 Update readme with new report path

📝 Fix decimals from upstream
