+++ stages/apis/coingecko/adapters.py
