+++ stages/apis/snapshot/queries.py
